Fix pretty print and iaa-wrapper way of doing the output
